Make sure that anything that uses lineWidth() do so after having rounded up its value.
The reasoning is that if, for example, you have nested msqrt's, then it
may happen that some lines may thicker/thinner than others even though
they all ought to be of the same width. This is because we use a
floating point based approach. So, to avoid this problem, we round up
the width of the line, thus ensuring that all lines will have the same
width.